Fixed one clang -Wfloat-conversion warning, variable type was just wrong

+++ testdir/minc2-float-voxel-test.c
@@ -138,7 +138,7 @@ main(void){
   midimhandle_t hdims[NDIMS];
   double outval;
   
-  double error_count = create_and_test_image("test-dbl.mnc", &hvol, hdims, outval);
+  int error_count = create_and_test_image("test-dbl.mnc", &hvol, hdims, outval);
 
   // clean up memory
   if(miclose_volume ( hvol )>0) error_count++;
